requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.60.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jwt==1.3.1
jwt_auth==0.4.3
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
urllib3==2.1.0
